requirements.txt
asgiref==3.8.1
Django==5.1.7
django-filter==25.1
djangorestframework==3.16.0
Markdown==3.7
python-decouple==3.8
sqlparse==0.5.3
